// apps/web/src/api/generated/models/PlatformSetting.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AnyJsonValue } from './AnyJsonValue';
export type PlatformSetting = {
    id?: string;
    settingKey?: string;
    settingValueJson?: AnyJsonValue;
    description?: string;
    updatedBy?: string | null;
    updatedAt?: string;
    createdAt?: string;
};